added balance logic to asset buying
With the new logic, a user may only buy an asset if they have enough funds to do so. In the case they lack it, they are shown an error message. When selling a holding, the amount will get added to the users, account in the function of the current price of that asset times the amoun they held.